+++ api/org.integratedmodelling.klab.api/src/org/integratedmodelling/klab/api/observations/IConfiguration.java
@@ -15,6 +15,8 @@
  */
 package org.integratedmodelling.klab.api.observations;
 
+import java.util.Collection;
+
 /**
  * The Interface IConfiguration.
  *
@@ -22,5 +24,13 @@ package org.integratedmodelling.klab.api.observations;
  * @version $Id: $Id
  */
 public interface IConfiguration extends IDirectObservation {
+
+	/**
+	 * A configuration is a detected pattern based on a set of observations. This
+	 * returns the observation that have made the configuration emerge.
+	 * 
+	 * @return the observations that define the configuration.
+	 */
+	Collection<IObservation> getTargetObservations();
 
 }