+++ klab.engine/src/main/java/org/integratedmodelling/klab/components/runtime/observations/Relationship.java
@@ -1,13 +1,13 @@
 package org.integratedmodelling.klab.components.runtime.observations;
 
 import org.integratedmodelling.klab.api.observations.IRelationship;
-import org.integratedmodelling.klab.components.runtime.RuntimeContext;
+import org.integratedmodelling.klab.engine.runtime.api.IRuntimeContext;
 import org.integratedmodelling.klab.owl.Observable;
 import org.integratedmodelling.klab.scale.Scale;
 
 public class Relationship extends CountableObservation implements IRelationship {
 
-	public Relationship(String name, Observable observable, Scale scale, RuntimeContext context) {
+	public Relationship(String name, Observable observable, Scale scale, IRuntimeContext context) {
 		super(name, observable, scale, context);
 	}